Fix Rate property on org.mpris.MediaPlayer2.Player
Rate should never be 0.0 (according to the spec), so we pick a sensible
default when we do not have a current track: 1.0.

Also, when the current item changes, the Rate may also change.  So add
it to the PropertiesChanged signal just to be sure.

Signed-off-by: Mirsal Ennaime <mirsal@videolan.org>
(cherry picked from commit 8f754d8715c72abc4ba1b0f26ae136c1db41437b)

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>