Fix Metadata marshalling when sending the PropertiesChanged signal
VLC was getting kicked from the D-Bus when a track was stopped, because
it was generating invalid data on the wire when sending the
PropertiesChanged signal for the Metadata property.

The issue was that if there was now no current track, GetInputMeta would
never be called and the the variant would never be populated with the
"a{sv}" structure that the call to dbus_message_iter_open_container
claimed it would be.

We now share the code that GetProperties used, which dealt with this
correctly, although now both use CurrentInput (which is what the signal
previously used) instead of CurrentPlayingItem (which is what
GetProperties previously used).

Signed-off-by: Mirsal Ennaime <mirsal@videolan.org>
(cherry picked from commit 5460e18e20719255d4ed1461594fe72530a19f85)

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>